storybook/requirements-api.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
stripe>=8.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
pydantic[email]>=2.0.0
